will not try to delete the default recording cell

+++ AirDrummer/AirDrummer/RecordingsTableViewController.swift
@@ -117,16 +117,18 @@ class RecordingsTableViewController: UITableViewController, AVAudioPlayerDelegat
     
     // Override to support editing the table view.
     override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCellEditingStyle, forRowAt indexPath: IndexPath) {
-        if editingStyle == .delete {
-            // Delete the row from the data source
-            let recordingURL = directoryContents.filter{ $0.deletingPathExtension().lastPathComponent == recordings[indexPath.row] }[0]
-            recordings.remove(at: indexPath.row)
-            try! FileManager.default.removeItem(at: recordingURL)
-            
-            tableView.deleteRows(at: [indexPath], with: .fade)
-        } else if editingStyle == .insert {
-            // Create a new instance of the appropriate class, insert it into the array, and add a new row to the table view
-        }    
+        if (recordings[indexPath.row] != "No Recordings Saved") {
+            if editingStyle == .delete {
+                // Delete the row from the data source
+                let recordingURL = directoryContents.filter{ $0.deletingPathExtension().lastPathComponent == recordings[indexPath.row] }[0]
+                recordings.remove(at: indexPath.row)
+                try! FileManager.default.removeItem(at: recordingURL)
+                
+                tableView.deleteRows(at: [indexPath], with: .fade)
+            } else if editingStyle == .insert {
+                // Create a new instance of the appropriate class, insert it into the array, and add a new row to the table view
+            }
+        }
     }
